Change setOutputPressure argument data type back to uint8_t from float. Same for M236 in Marlin_main.cpp (related). Something about the float cause a linker error.

+++ Marlin/Marlin_main.cpp
@@ -4804,7 +4804,7 @@ inline void gcode_M226() {
    */
   inline void gcode_M236() {
     if(code_seen('S')) {
-      float psi = code_value();
+      uint8_t psi = code_value();
       // Check that desired pressure is within range allowed
       if((psi <= OUTPUT_PSI_MAX) && (psi >= OUTPUT_PSI_MIN)) {
         setOutputPressure(psi);
