+++ Marlin/Regulator.cpp
@@ -7,7 +7,7 @@
 #include "MCP4725.h"
 
 // Set the output pressure (in psi) of the regulator.
-void setOutputPressure(float pressure) {
+void setOutputPressure(uint8_t pressure) {
 	// 4095/(130.355 psi)
 	uint16_t digital_val = (uint16_t)(31.414215 * pressure);
 	// Write value to DAC
